--- src/main/java/domain/lotto/LottoTicketGenerator.java
@@ -1,5 +1,0 @@
-package domain.lotto;
-
-public interface LottoTicketGenerator {
-    LottoTicket generate();
-}
