+++ src/main/java/domain/lotto/LottoWinningNumber.java
@@ -1,5 +1,6 @@
 package domain.lotto;
 
+import domain.lotto.ticket.LottoTicket;
 import util.ListUtils;
 
 import java.util.Collections;
