+++ src/main/java/domain/lotto/prize/judgment/impl/FirstLottoPrizeJudgment.java
@@ -1,20 +1,20 @@
-package domain.judgment.impl;
+package domain.lotto.prize.judgment.impl;
 
-import domain.judgment.LottoPrizeJudgment;
 import domain.lotto.LottoNumber;
-import domain.lotto.LottoTicket;
 import domain.lotto.LottoWinningNumber;
-import util.ListUtils;
+import domain.lotto.prize.judgment.LottoPrizeJudgment;
+import domain.lotto.ticket.LottoTicket;
 
 import java.util.List;
 
-public class FourthLottoPrizeJudgment implements LottoPrizeJudgment {
+public class FirstLottoPrizeJudgment implements LottoPrizeJudgment {
 
     @Override
     public boolean judge(LottoTicket lottoTicket, LottoWinningNumber lottoWinningNumber) {
         List<LottoNumber> lottoNumbers = lottoTicket.getLottoNumbers();
         List<LottoNumber> winningNumbers = lottoWinningNumber.getLottoNumbers();
 
-        return ListUtils.countCommonElements(lottoNumbers, winningNumbers) == 4;
+        return lottoNumbers.containsAll(winningNumbers);
     }
+
 }
