+++ src/main/java/domain/lotto/prize/judgment/impl/NoneLottoPrizeJudgment.java
@@ -1,20 +1,20 @@
-package domain.judgment.impl;
+package domain.lotto.prize.judgment.impl;
 
-import domain.judgment.LottoPrizeJudgment;
 import domain.lotto.LottoNumber;
-import domain.lotto.LottoTicket;
 import domain.lotto.LottoWinningNumber;
+import domain.lotto.prize.judgment.LottoPrizeJudgment;
+import domain.lotto.ticket.LottoTicket;
 import util.ListUtils;
 
 import java.util.List;
 
-public class ThirdLottoPrizeJudgment implements LottoPrizeJudgment {
+public class NoneLottoPrizeJudgment implements LottoPrizeJudgment {
 
     @Override
     public boolean judge(LottoTicket lottoTicket, LottoWinningNumber lottoWinningNumber) {
         List<LottoNumber> lottoNumbers = lottoTicket.getLottoNumbers();
         List<LottoNumber> winningNumbers = lottoWinningNumber.getLottoNumbers();
 
-        return ListUtils.countCommonElements(lottoNumbers, winningNumbers) == 5;
+        return ListUtils.countCommonElements(lottoNumbers, winningNumbers) <= 2;
     }
 }
