+++ src/main/java/domain/lotto/ticket/generator/LottoTicketGenerator.java
@@ -1,0 +1,7 @@
+package domain.lotto.ticket.generator;
+
+import domain.lotto.ticket.LottoTicket;
+
+public interface LottoTicketGenerator {
+    LottoTicket generate();
+}
