+++ src/test/java/domain/lotto/LottoPrizeTest.java
@@ -1,5 +1,7 @@
 package domain.lotto;
 
+import domain.lotto.prize.LottoPrize;
+import domain.lotto.ticket.LottoTicket;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
