+++ src/test/java/domain/lotto/LottoTicketTest.java
@@ -1,5 +1,6 @@
 package domain.lotto;
 
+import domain.lotto.ticket.LottoTicket;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
